Fix regression broken Stack + add DOS line ending support
The Cell was deleting itself in cascade which was breaking the right
behavior of the Stack. The BaseList is now responsible for this cascade
deletion feature.

DOS (\r\n) line ending wasn't taken into account in files which was
breaking string interpretation. Now supported. Using the Stack class
when reading file content simplified the reading algorithm by a lot. All
is now more robust when parsing statistical data.